requirements.txt
boto3
pandas
scikit-learn
requests
kafka-python
websocket-client